Refactor: extract shared types, move Scrubber and Spectrogram into folders with custom hooks
- Extract TrackId, TrackColor, Track types from projectPageReducer into
  src/types/track.ts so hooks, signals, and services no longer import
  from components
- Move Scrubber into scrubber/ folder, extract useScrubber hook for
  scroll sync, playback animation, user interaction, and timeline scaling
- Move Spectrogram into spectrogram/ folder, extract useSpectrogramCache
  hook for async cache loading and spectrogramRenderer for drawing utils
- Update all imports across the codebase (25+ files)
