+++ src/components/workstation/Channel.tsx
@@ -5,7 +5,7 @@ import {
 } from '@ant-design/icons';
 import { Button, Slider } from 'antd';
 import classNames from 'classnames';
-import { Track } from '../project/projectPageReducer';
+import { type Track } from '../../types/track';
 import './Channel.css';
 import { useChannelControls } from './useChannelControls';
