+++ src/components/workstation/Workstation.tsx
@@ -6,10 +6,10 @@ import { useTransportBridge } from '../../hooks/useTransportBridge';
 import { pixelsPerSecond as pixelsPerSecondSignal } from '../../signals/workstationSignals';
 import Dropzone from '../dropzone/Dropzone';
 import { useFileDropzone } from '../dropzone/useFileDropzone';
-import { Track } from '../project/projectPageReducer';
+import { type Track } from '../../types/track';
 import EmptyTimeline from './EmptyTimeline';
 import Mixer from './Mixer';
-import Scrubber from './Scrubber';
+import Scrubber from './scrubber/Scrubber';
 import Timeline from './Timeline';
 import Toolbar from './Toolbar';
 import './Workstation.css';
