+++ src/components/workstation/spectrogram/Spectrogram.css
