+++ src/components/workstation/spectrogram/Spectrogram.tsx
@@ -1,11 +1,12 @@
-import { useEffect, useRef, useState } from 'react';
-import { useAnimationFrame } from '../../hooks/useAnimationFrame';
-import { useAudioService } from '../../hooks/useAudioService';
-import { useTrackVolume } from '../../hooks/useTrackVolume';
-import { TrackSpectrogramEntry } from '../../services/SpectrogramCache';
-import { isPlaying, transportTime } from '../../signals/transportSignals';
-import { Track, TrackColor } from '../project/projectPageReducer';
+import { useRef } from 'react';
+import { useAnimationFrame } from '../../../hooks/useAnimationFrame';
+import { useAudioService } from '../../../hooks/useAudioService';
+import { useTrackVolume } from '../../../hooks/useTrackVolume';
+import { isPlaying, transportTime } from '../../../signals/transportSignals';
+import { type Track } from '../../../types/track';
 import './Spectrogram.css';
+import { drawLiveColumn } from './spectrogramRenderer';
+import { useSpectrogramCache } from './useSpectrogramCache';
 
 type SpectrogramProps = {
   height: number;
@@ -15,12 +16,6 @@ type SpectrogramProps = {
 
 const TILE_WIDTH = 4096;
 const SCROLL_CONTAINER_CLASS = '.scrubber__timeline';
-const LIVE_COLUMN_WIDTH = 2;
-
-// Match the OfflineAnalyser's AnalyserNode dB range for visual consistency
-const MIN_DB = -80;
-const MAX_DB = -30;
-const DB_RANGE = MAX_DB - MIN_DB;
 
 const Spectrogram = ({ height, pixelsPerSecond, track }: SpectrogramProps) => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
@@ -32,30 +27,7 @@ const Spectrogram = ({ height, pixelsPerSecond, track }: SpectrogramProps) => {
   const audioService = useAudioService();
   const audioBuffer = audioService.retrieveAudioBuffer(trackId);
 
-  const [entry, setEntry] = useState<TrackSpectrogramEntry | undefined>();
-
-  // Trigger cache analysis on mount if not already cached
-  useEffect(() => {
-    if (!audioBuffer) return;
-
-    const cached = audioService.spectrogramCache.getEntry(trackId);
-    if (cached) {
-      setEntry(cached);
-      return;
-    }
-
-    let cancelled = false;
-    audioService.spectrogramCache
-      .analyse(trackId, audioBuffer, color)
-      .then(() => {
-        if (!cancelled) {
-          setEntry(audioService.spectrogramCache.getEntry(trackId));
-        }
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [trackId, audioBuffer, color, audioService]);
+  const entry = useSpectrogramCache(trackId, audioBuffer, color);
 
   const duration = audioBuffer?.duration ?? 0;
   const containerWidth = duration * pixelsPerSecond;
@@ -159,56 +131,5 @@ const Spectrogram = ({ height, pixelsPerSecond, track }: SpectrogramProps) => {
     </div>
   );
 };
-
-/**
- * Draws a single bright column on the canvas at the playhead position,
- * reflecting live post-effects frequency content during playback.
- *
- * Frequency bins are grouped into pixel rows (max intensity per row)
- * and rendered with additive compositing so the overlay appears brighter
- * than the static spectrogram tiles underneath.
- */
-function drawLiveColumn(
-  ctx: CanvasRenderingContext2D,
-  frequencyData: Float32Array,
-  playheadX: number,
-  height: number,
-  color: TrackColor,
-): void {
-  if (playheadX < -LIVE_COLUMN_WIDTH || playheadX > ctx.canvas.width) return;
-
-  const bins = frequencyData.length;
-  const binsPerRow = bins / height;
-  const { r, g, b } = color;
-  const x = Math.round(playheadX);
-
-  ctx.save();
-  ctx.globalCompositeOperation = 'lighter';
-
-  for (let row = 0; row < height; row++) {
-    const startBin = Math.floor(row * binsPerRow);
-    const endBin = Math.floor((row + 1) * binsPerRow);
-    let maxIntensity = 0;
-    for (let bin = startBin; bin < endBin; bin++) {
-      const intensity = dbToByte(frequencyData[bin]);
-      if (intensity > maxIntensity) maxIntensity = intensity;
-    }
-    if (maxIntensity === 0) continue;
-
-    const alpha = maxIntensity / 255;
-    // bin 0 → bottom row, last bin → top row
-    const y = height - row - 1;
-    ctx.fillStyle = `rgba(${r}, ${g}, ${b}, ${alpha})`;
-    ctx.fillRect(x, y, LIVE_COLUMN_WIDTH, 1);
-  }
-
-  ctx.restore();
-}
-
-function dbToByte(db: number): number {
-  if (db <= MIN_DB) return 0;
-  if (db >= MAX_DB) return 255;
-  return Math.round(((db - MIN_DB) / DB_RANGE) * 255);
-}
 
 export default Spectrogram;
