+++ src/types/track.ts
@@ -1,0 +1,14 @@
+export type TrackId = string;
+
+export type TrackColor = {
+  r: number;
+  g: number;
+  b: number;
+};
+
+export type Track = {
+  trackId: TrackId;
+  color: TrackColor;
+  fileName: string;
+  index: number;
+};
